# UserService/UserServiceApi/src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: User Service API
  description: Sample User Service for Opsec
  version: 0.0.1
  
servers:
  - url: http://localhost:8000
    description: local server
  
paths:
  /user/read/{userId}:
    get:
      summary: Returns a user for given Id.
      operationId: readUser
      
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type : string
          
      responses:
        400:
          description: Bad request
        200:
          description: successful response
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/UserForm'
                
  /user/update:
    post:
      summary: Update user details
      operationId: updateUser
      
      requestBody:
        description: consists of updated user details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserForm'
          
      responses:
        400:
          description: Bad request
        200:
          description: successful response
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/updateUserResponse'
                
  /user/create:
    post:
      summary: create a new user
      operationId: createUser
      
      requestBody:
        description: consists of updated user details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserForm'
          
      responses:
        400:
          description: Bad request
        200:
          description: successful response
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/createUserResponse'
                
  /user/delete/{userId}:
    post:
      summary: Delete a user with given id
      operationId: deleteUser
      
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type : string
          
      responses:
        400:
          description: Bad request
        200:
          description: successful response
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/deleteUserResponse'
        
components:
  schemas:
    UserForm:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dob:
          type: string
          pattern: '[0-9]{2}-[0-9]{2}-[0-9]{4}'
        title:
          type: string

    CreateUserForm:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dob:
          type: string
          pattern: '[0-9]{2}-[0-9]{2}-[0-9]{4}'
        title:
          type: string
    createUserResponse:
      type: object
      properties:
        uniqueId:
          type: string
    
    updateUserResponse:
      type: object
      properties:
        uniqueId:
          type: string
        updated:
          type: boolean
      
    deleteUserResponse:
      type: object
      properties:
        deleted:
          type: boolean 